enhance: add update functionality when delete or editing income or expense, and fix some errors
