+++ client/components/RecentActivity.js
@@ -2,34 +2,20 @@ import {
   View,
   FlatList,
 } from "react-native";
-import React, { useMemo } from "react";
-import useExpensesAndIncomes from "../hooks/useExpenseAndIncomes";
+import React from "react";
 import CustomText from "./CustomText";
 import ActivityDisplay from "./ActivityDisplay";
 import { recentActivity } from "../styles/components/recent-activity";
 import { homeStyles } from "../styles/screens/home";
 
-//Function to combine and sort data
-const mergeAndSortData = (incomes, expenses) => {
-  return [...incomes, ...expenses]
-    .map((item, index) => ({ ...item, id: (index + 1).toString() }))
-    .sort((a, b) => new Date(b.date) - new Date(a.date))
-    .slice(0, 10);
-};
 
-export default function RecentActivity({ onPress }) {
-  const {expenses, incomes, loading} = useExpensesAndIncomes();
-  const combinedData = useMemo(
-    () => mergeAndSortData(incomes, expenses),
-    []
-  );
-
+export default function RecentActivity({ onPress, activity, loading }) {
   return (
     <View style={recentActivity.container}>
       <View style={recentActivity.header}>
         <CustomText text="Actividad Reciente" type="TitleSmall" />
       </View>
-      {loading 
+      {loading || activity.length === 0 
         ? (
           <View style={homeStyles.recentActivityContainer}>
             <CustomText 
@@ -40,7 +26,7 @@ export default function RecentActivity({ onPress }) {
         ): (
           <FlatList
           showsVerticalScrollIndicator={false}
-          data={combinedData}
+          data={activity}
           keyExtractor={(item) => item.id}
           renderItem={({ item }) => (
             <ActivityDisplay
